Do not assume http_response_header set during exception
The variable $http_response_header may not be set during an exception. A
typical use case is calling out to the Twitter API during Twitter
authentication when the callback URI has not been registered with
Twitter. Because of how the Twitter API responds $http_response_header
is not set and during the handling of the exception the missing
$http_response_header masks the real issue. This commit changes the code
to not assume that $http_response_header is set.